More accurate summary by parsing source code
This is an alternative way to grab wikipedia summary which is more accurate and uses curl, grep and sed.

Explanation:
It grabs the source code of a Wikipedia article using curl and keeps it in a temporary file $TMPFILE, 
grep the appropriate summary lines, uses sed to remove a last unimportant line (this should probably be replaced with a less cumbersome command), 
and format-print it as usual and ${BROWSER} -dump it. 
After that it removes the temporary file.

Possible TODO: 
Add a test for the existence of curl. Allow wget as an alternative.